pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdfhikaku"
version = "0.1.0"
description = "PDF comparison application with text diff visualization"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.47.1",
    "pymupdf>=1.23.0",
    "pdfminer.six>=20231228",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "rapidfuzz>=3.6.0",
    "python-Levenshtein>=0.20.0",
    "regex>=2023.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "matplotlib>=3.7.0",
    "reportlab>=4.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
    "psutil>=5.9.0",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.0",
    "scipy>=1.10.0"
]

[project.optional-dependencies]
pdf = ["weasyprint>=60.0", "pdfkit>=1.0.0"]
openai = ["openai>=1.0.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["pdfhikaku*"]
